Don't expose platform::util at top level
This moves various utility methods into `Application`, following the
new wrapper pattern.

It also cleans up imports a bit, because I have ADHD
