+++ druid-shell/src/platform/windows/menu.rs
@@ -21,8 +21,8 @@ use winapi::shared::basetsd::*;
 use winapi::shared::windef::*;
 use winapi::um::winuser::*;
 
+use super::util::ToWide;
 use crate::hotkey::HotKey;
-use crate::util::ToWide;
 
 /// A menu object, which can be either a top-level menubar or a
 /// submenu.
