+++ druid-shell/src/platform/windows/paint.rs
@@ -39,7 +39,7 @@ use direct2d;
 use direct2d::enums::{AlphaMode, RenderTargetType};
 use direct2d::render_target::{DxgiSurfaceRenderTarget, GenericRenderTarget, HwndRenderTarget};
 
-use crate::util::as_result;
+use super::util::as_result;
 use crate::Error;
 
 /// Context for painting by app into window.
