+++ druid/src/app.rs
@@ -19,8 +19,9 @@ use std::rc::Rc;
 use std::sync::Arc;
 
 use crate::kurbo::Size;
+use crate::shell::application::Application;
 use crate::shell::window::WindowHandle;
-use crate::shell::{init, runloop, Error as PlatformError, WindowBuilder};
+use crate::shell::{runloop, Error as PlatformError, WindowBuilder};
 use crate::win_handler::AppState;
 use crate::window::{Window, WindowId};
 use crate::{theme, AppDelegate, Data, DruidHandler, Env, LocalizedString, MenuDesc, Widget};
@@ -94,7 +95,7 @@ impl<T: Data + 'static> AppLauncher<T> {
     /// Returns an error if a window cannot be instantiated. This is usually
     /// a fatal error.
     pub fn launch(mut self, data: T) -> Result<(), PlatformError> {
-        init();
+        Application::init();
         let mut main_loop = runloop::RunLoop::new();
         let mut env = theme::init();
         if let Some(f) = self.env_setup.take() {
